+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ Release 0.4 UNRELEASED
 ----------------------
 * fixed some packaging bugs.
 * made the data model more tolerant to missing symbols.
+* fixed symbol resolving bug.
 
 Release 0.3 16/04/2013
 -----------------------