+++ CHANGELOG.rst
@@ -1,4 +1,4 @@
-Release 0.4 UNRELEASED
+Release 0.4 19/04/2013
 ----------------------
 * fixed some packaging bugs.
 * made the data model more tolerant to missing symbols.